+++ status.js
@@ -4,14 +4,21 @@ const Joi = require('joi');
 const Wrap = require('./wrap');
 
 /**
- * A dynamo model that is used for storing the general status for
+ * Returns a Wrapped Dynastar model that is used for storing the general status for
  * a given pkg, env, version. This is used for computing progress in
  * conjunction with StatusCounter using the `total` field
+ *
+ * @function status
+ * @param {Object} dynamo Dynamo object model
+ * @returns {Wrap} Status
  */
 module.exports = function status(dynamo) {
   const hashKey = 'key';
   const model = dynamo.define('status', {
     hashKey,
+    timestamps: true,
+    createdAt: 'createDate',
+    updatedAt: 'updateDate',
     tableName: 'status',
     schema: {
       key: Joi.string(),
@@ -21,8 +28,6 @@ module.exports = function status(dynamo) {
       previousVersion: Joi.string(),
       total: Joi.number(),
       error: Joi.boolean(),
-      createDate: Joi.date(),
-      updateDate: Joi.date(),
       complete: Joi.boolean()
     }
   });