+++ test/unit.test.js
@@ -1,30 +1,22 @@
 const assume = require('assume');
 const ModelWrap = require('../wrap');
-const models = require('..');
-
-const { mocks, helpers } = require('datastar-test-tools');
+const dynamodb = require('dynamodb');
+const models = require('..')(dynamodb);
 
 describe('warehouse.ai-status-models (unit)', function () {
-  let dal;
-
-  beforeEach(function () {
-    const datastar = helpers.connectDatastar({ mock: true }, mocks.datastar());
-    dal = models(datastar);
-  });
-
   it('should return the appropriate object', function () {
-    assume(dal.Status).is.instanceof(ModelWrap);
-    assume(dal.StatusCounter).is.instanceof(ModelWrap);
-    assume(dal.StatusEvent).is.instanceof(ModelWrap);
-    assume(dal.StatusHead).is.instanceof(ModelWrap);
+    assume(models.Status).is.instanceof(ModelWrap);
+    assume(models.StatusCounter).is.instanceof(ModelWrap);
+    assume(models.StatusEvent).is.instanceof(ModelWrap);
+    assume(models.StatusHead).is.instanceof(ModelWrap);
   });
 
   it('should have an ensure and drop functions', function () {
-    assume(dal.ensure).is.a('function');
-    assume(dal.drop).is.a('function');
+    assume(models.ensure).is.a('function');
+    assume(models.drop).is.a('function');
   });
 
-  it('StatusCounter model should have increment function', function () {
-    assume(dal.StatusCounter.increment).is.an('asyncfunction');
+  it('StatusCounter model should have increment and decrement function', function () {
+    assume(models.StatusCounter.increment).is.an('asyncfunction');
   });
 });