+++ wrap.js
@@ -95,14 +95,6 @@ class Wrap {
   drop() {
     return thenify(this.model, 'dropTables');
   }
-  /**
-   * Get the raw cassandra-driver because we need to do some special shit for
-   * counters
-   * @returns {Thenable} raw cassandra driver
-   */
-  _getConnection() {
-    return thenify(this.model.connection, 'getConnectionPool', null, false);
-  }
 }
 
 module.exports = Wrap;